% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cta-client.R
\name{ctaclient}
\alias{ctaclient}
\title{R6 class for accessing the CTA Transit Tracker API}
\description{
A client to access the CTA Transit Tracker API. Can make requests to the API
as long as the user has their API Key set in their R environ as \code{TRAIN_KEY}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_url}}{The base URL for the api. Generally shouldn't need to change,
but leaving room for accessing future CTA APIs.}

\item{\code{last_request}}{The last API request set up via \code{request()}}

\item{\code{last_response}}{The most recent response from the API, if one has
been made since initialization.}

\item{\code{last_response_body}}{The body of the most recent response from the API,
typically as list.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ctaClient-new}{\code{ctaclient$new()}}
\item \href{#method-ctaClient-request}{\code{ctaclient$request()}}
\item \href{#method-ctaClient-send}{\code{ctaclient$send()}}
\item \href{#method-ctaClient-print}{\code{ctaclient$print()}}
\item \href{#method-ctaClient-clone}{\code{ctaclient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ctaClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-ctaClient-new}{}}}
\subsection{Method \code{new()}}{
Set up a new api client. Can change the base url here if needed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctaclient$new(api_url = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_url}}{The base url for the api. Will default to
"http://lapi.transitchicago.com/api/1.0" if none provided. Left for
future expansion.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ctaClient-request"></a>}}
\if{latex}{\out{\hypertarget{method-ctaClient-request}{}}}
\subsection{Method \code{request()}}{
Build a \code{httr2} request object, but don't send it just yet.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctaclient$request(verb, path, body = NULL, query = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verb}}{Method for request (e.g. 'GET', 'POST', ...)}

\item{\code{path}}{Either character string or vector defining the api path appended
to the end of the base url.}

\item{\code{body}}{A list of any body contents necessary for the api request.}

\item{\code{query}}{A list of any additional query parameters to append to the
api request.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ctaClient-send"></a>}}
\if{latex}{\out{\hypertarget{method-ctaClient-send}{}}}
\subsection{Method \code{send()}}{
Actually perform the request set up by \code{request()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctaclient$send()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ctaClient-print"></a>}}
\if{latex}{\out{\hypertarget{method-ctaClient-print}{}}}
\subsection{Method \code{print()}}{
Print some basic information about the client.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctaclient$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ctaClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ctaClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctaclient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
