% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{route}
\alias{route}
\title{R6 Class Representing a CTA Train Route}
\description{
This class can be used to grab the positions of all trains along a
specific CTA Train Line (Brown, Red, Blue, etc.).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{client}}{Instance of the ctaclient class.}

\item{\code{trains}}{A list of trains on the route with additional information.}

\item{\code{time_stamp}}{Time stamp (in US/Central time) of the most recent API
call to retrieve \code{trains} information.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{rt_name}}{(\code{character}) Name of the route set at initialization.}

\item{\code{rt_code}}{(\code{character}) Special code based on \code{rt_name}. Note that
sometimes this value will be exactly equal to \code{rt_name}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-route-new}{\code{route$new()}}
\item \href{#method-route-print}{\code{route$print()}}
\item \href{#method-route-get_positions}{\code{route$get_positions()}}
\item \href{#method-route-clone}{\code{route$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-route-new"></a>}}
\if{latex}{\out{\hypertarget{method-route-new}{}}}
\subsection{Method \code{new()}}{
Generates a new \code{route} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{route$new(rt_name, client = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rt_name}}{Name of the desired route (e.g. "Brown", "Red", ...). See
\code{ctaTracker::train_routes} for full list of options.}

\item{\code{client}}{Instance of ctaclient API client to interact with the
CTA Train Tracker API. Will default to basic client if none provided.
Other values generally only needed for testing purposes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-route-print"></a>}}
\if{latex}{\out{\hypertarget{method-route-print}{}}}
\subsection{Method \code{print()}}{
Print some basic information about the route.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{route$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-route-get_positions"></a>}}
\if{latex}{\out{\hypertarget{method-route-get_positions}{}}}
\subsection{Method \code{get_positions()}}{
Grab all current trains on the route and any information about them.
Send results to \code{self$trains}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{route$get_positions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-route-clone"></a>}}
\if{latex}{\out{\hypertarget{method-route-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{route$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
